Validation

+++ Controllers/FilesController.cs
@@ -1,6 +1,8 @@
 ﻿using EcoLease_API.Services;
+using EcoLease_API.Validators;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using FluentValidation.Results;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
@@ -24,10 +26,18 @@ namespace EcoLease_API.Controllers
             _fileService = fileService;
         }
 
+        FileValidator validator = new FileValidator();
+        
+
         // GET api/Files/{img1.jpg}
         [HttpGet("{fileName}")]
         public IActionResult GetFile(string fileName)
         {
+            FluentValidation.Results.ValidationResult validation = validator.Validate(fileName);
+            if (!validation.IsValid)
+            {
+                return BadRequest("Wrong Parameters!");
+            }
 
             try
             {
@@ -44,6 +54,12 @@ namespace EcoLease_API.Controllers
         [HttpPost("upload")]
         public IActionResult UploadImage(IFormFile file)
         {
+            FluentValidation.Results.ValidationResult validation = validator.Validate(file.FileName);
+            if (!validation.IsValid)
+            {
+                return BadRequest("Wrong Parameters!");
+            }
+
             try
             {
                 if (_fileService.UploadFile(file))
@@ -62,6 +78,12 @@ namespace EcoLease_API.Controllers
         [HttpDelete("{fileName}")]
         public IActionResult RemoveFile(string fileName)
         {
+            FluentValidation.Results.ValidationResult validation = validator.Validate(fileName);
+            if (!validation.IsValid)
+            {
+                return BadRequest("Wrong Parameters!");
+            }
+
             try
             {
                 if (_fileService.RemoveFile(fileName))
